fix: MOE整改风险接口 GetAttrPointer<int>()

+++ moe/moe_token_permute_grad/op_host/moe_token_permute_grad_tiling.cpp
@@ -324,7 +324,7 @@ ge::graphStatus PermuteTilingCompute(gert::TilingContext *context, const int64_t
 
 static ge::graphStatus Tiling4MoeTokenPermuteGrad(gert::TilingContext *context)
 {
-    const int *top_k = context->GetAttrs()->GetAttrPointer<int>(0);
+    const int64_t *top_k = context->GetAttrs()->GetAttrPointer<int64_t>(0);
     int64_t topk = static_cast<int64_t>(*top_k);
     return PermuteTilingCompute(context, topk);
 }
